requirements.txt
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.2.0
importlib_resources==6.4.5
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
zipp==3.20.2
